Complete Rework
This is a complete overhaul that adds tests and removes a lot of the
annoyances from previous versions. This is in beta stage, and drops
assassination support temporarily.

Major features include:

Profiles - It is very easy to work on your own custom profile without
overriding the default one, as well as set up different profiles for
raiding as opposed to questing or even dungeons.

No more reloading - It is no longer necessary to reload when adding new
suggestions.

Tests - This is setup with tests in mind, as well as a fairly large base
of tests. This should theoretically result in fewer errors.

Currently missing from this version is the ability to detect when spells
were applied (and their corresponding multiplier), which was key to
assassination and is the main reason for its temporary removal. This
should be fixed in the upcoming version.

Crimson vial is not in this version, it will be in the upcoming version.

For those of you testing this version, please include your
SavedVariables file in bug reports. It is accessed in World of
Warcraft/WTF/Account/<your account number, usually only 1 option>/<your
realm>/<character name>/Saved Variables/Torpedo.lua