Skada integration
This update includes support for skada integration, which allows
checking for time-to-kill solo and time-to-kill raid. This uses skada
dps for the current fight if there is one, and all the saved skada data
otherwise.

Theoretically, it gracefully doesn't check these if Skada isn't
installed.

This update all paves the way for checking bleed multipliers, which is
essential for assassination.