Add assassination back
This update adds assassination back in and a few other convienience
functions like cloning profiles.

The rotation is based on icy-veins, but as always that rotation does not
feel or perform very well when translated word-for-word, especially if
you get any crits or bonus combo points. This rotation is not ideal, but
it should be comparable to the previous release version and the other
changes make this update more than worth it for assassination rogues.

I need assassination rogues with the golden artifact traits to test
those, since they change several spell ids. However, the old release
version doesn't magically handle those spells either, so it's still
worth an update. Thus, this is marked as release!