Bug fixes, config improvements
Add options to change the order of skills inside a specialization. This
is purely a visual effect.

Bonus points: There is a link to setup for subtlety in the description
if you choose Weaponmaster, Alacrity, and Death from Above for a fun aoe
build. Pulls around 240k single target from the front in my tests which
is a 15-25k dps loss over my normal build, which is Master of Subtlety,
Premeditation and Master of Shadows, but scales better if you have 2+
targets. Plus, it's really fun to jump around a bunch and have insane
haste. If your alacrity stacks ever wear off (20 seconds of no combo
finishers), it's completely not worth it since it takes almost a full
minute to ramp up.